+++ src/generic_core/remote-connection.spec.ts
@@ -10,7 +10,7 @@
 import freedomMocker = require('../../../third_party/uproxy-lib/freedom/mocks/mock-freedom-in-module-env');
 import freedom_mocks = require('../mocks/freedom-mocks');
 freedom = freedomMocker.makeMockFreedomInModuleEnv({
-    'core.storage': () => { return new freedom_mocks.MockStorage(); },
+    'core.storage': () => { return new freedom_mocks.MockFreedomStorage(); },
     'loggingcontroller': () => { return new freedom_mocks.MockLoggingController(); },
     'metrics': () => { return new freedom_mocks.MockMetrics(); }
 });
